sysfs: don't pull modules to sysinit runlevel
Commit 73cdf10f1f51 ("Deprecate automatic loading of modules") introduced
"want modules" dependency for the "sysfs" service.

By default the "sysfs" service is installed in the "sysinit" runlevel, is
one of the very first services to start on boot (often it is the first one)
and it is always started before udev since udev depends on it.
The "modules" service on the other hand is supposed to start a bit later,
in the "boot" runlevel.

However, since that mentioned earlier commit the "modules" service will be
pulled in to the "sysinit" runlevel because a "want" dependency will do it
regardless in which runlevel this service is actually installed.
This breaks things like USB modem mode switching.

Change this dependency to an "use" which does not automatically pull in the
"modules" service into the "sysinit" runlevel and inform an user that he
needs to add it manually to this runlevel in case that he boots an EFI
Linux system AND efivarfs support wasn't built in into the kernel AND
efivarfs module wasn't loaded already by an initrd.