Add file path to error messages
When accessing a file fails, printing out an error message is generally a
good idea, but omitting the name of the file you tried to access makes it
utterly useless.

For example: "fopen: no such file or directory" is entirely meaningless.
We don't know what file the program tried to access and have no clue
what we should fix. We need a tool like strace to find out what exactely
happened. This kind of error message is about as bad as no error
message at all.

This patch tries to fix a number of obvious cases where perror or
strerror(errno) is used to report failure caused by a filesystem or
file handle access, but the file name was missing.

Signed-off-by: David Oberhollenzer <david.oberhollenzer@tele2.at>