fix alternative conf and init dir support
Previously, INITDIR and CONFDIR could be set from the command line.
However, they only affected where things were installed and didn't
update the runtime search paths.

Update to allow the directory names to be specified and restricted to
SYSCONFDIR.

RC_PKG_ and RC_LOCAL_ are left with the standard names as presumably
there wouldn't be conflicts in non-system paths.