src/test/runtests.sh: drop 'readelf'-based tests
The 'readelf'-based tests cover a few situations:
1. undefined symbols in shared libraries
2. unexpected exports in shared libraries

Bug #575958 shows that [2.] implementation is too simplistic
in assuming that presence of relocation equals to export presence.

It is incorrect for PLT stubs and local symbols.
Let's just drop these tests.

If one neds to cover [1.] it is better to use LDFLAGS=-Wl,--no-undefined.

Reported-by: Benda Xu
Bug: https://bugs.gentoo.org/575958
Signed-off-by: Sergei Trofimovich <slyfox@gentoo.org>