net-online: fix get_interfaces() and add minimum_up
This change incorporates a fork that I made for k3OS that helps
solve a number of network availability race-conditions.

See rancher/k3os#184.

~This fixes #316 by removing the `test -h $ifname` check altogether.
Not only was the check backwards, resulting in undetected interfaces,
but on k3OS at least, everything in /sys/class/net is a symlink.~

This sidesteps "file not found" errors when attempting to read
carrier and operstate for explicitly specified interfaces that do not
exist.

Lastly, a new `conf.d` variable, `minimum_up`, allows for net-online to
return successfully when at least that number of interfaces are in the
"up" state. The default empty values results in the old behavior of
expecting all specified (or detected) interfaces to be "up".

Fixes #316
Supercedes #317